+++ straxen/storage/__init__.py
@@ -1,0 +1,14 @@
+from . import online_monitor_frontend
+from .online_monitor_frontend import *
+
+from . import rucio_remote
+from .rucio_remote import *
+
+from .rucio_local import *
+from . import rucio_local
+
+from . import rundb
+from .rundb import *
+
+from . import mongo_storage
+from .mongo_storage import *